# Separate CArray and family from `core.pyx`