Similar refactoring to #2785
Removes `include 'carray.pxi'` from `core.pyx`.

CPU performance benchmark
#### Before
```
empty               :     0.233 us   +/- 0.177 (min:    0.188) us
sum                 :    17.517 us   +/- 1.471 (min:   16.689) us
sum_huge            :    21.874 us   +/- 4.227 (min:   18.048) us
add                 :    12.300 us   +/- 1.122 (min:   11.752) us
add_out             :     9.184 us   +/- 1.086 (min:    8.735) us
add_out_huge        :    11.666 us   +/- 2.149 (min:   10.163) us
userkernel_create   :     2.798 us   +/- 0.667 (min:    2.630) us
userkenel           :     8.929 us   +/- 0.933 (min:    8.561) us
userkernel_huge     :    10.652 us   +/- 1.015 (min:    9.890) us
```

#### This PR
```
empty               :     0.213 us   +/- 0.013 (min:    0.191) us
sum                 :    17.852 us   +/- 1.594 (min:   16.859) us
sum_huge            :    21.445 us   +/- 2.672 (min:   19.947) us
add                 :    12.286 us   +/- 1.482 (min:   11.742) us
add_out             :     9.317 us   +/- 1.137 (min:    8.825) us
add_out_huge        :    11.169 us   +/- 1.502 (min:   10.355) us
userkernel_create   :     2.887 us   +/- 0.642 (min:    2.694) us
userkenel           :     9.054 us   +/- 1.217 (min:    8.680) us
userkernel_huge     :    10.510 us   +/- 1.224 (min:    9.784) us
```